refactor - move `Base*` classes into `base` module: * rename `SingleSpider` -> `NewsArticleSpider`
