+++ scrapy_ntk/__init__.py
@@ -2,6 +2,6 @@ from .parsing import *
 from .exporting import *
 from .tools import *
 
-from .spider import SingleSpider
+from .spider import NewsArticleSpider
 from .item import ArticleItem, FIELDS
 from .downloader import ProxyManagerDM
