+++ scrapy_ntk/pipeline.py
@@ -1,6 +1,6 @@
-import abc
 import logging
 
+from .base import BaseArticlePipeline
 from .config import cfg
 from .exporting import (
     GSpreadMaster,
@@ -12,9 +12,7 @@ from .exporting import (
     GSpreadRow,
     BackupGSpreadRow,
 )
-from .item import ArticleItem
-from .spider import BaseSpider
-from .exporting.base import BaseArticleItemExporter
+from .spider import BaseArticleSpider
 
 
 logger = logging.getLogger(__name__)
@@ -36,105 +34,9 @@ ENABLE_SHUB = _to_boolean(cfg.enable_shub)
 ENABLE_DATABASE = _to_boolean(cfg.enable_database)
 
 
-class ArticlePipeline(abc.ABC):
+class GSpreadPipeline(BaseArticlePipeline):
 
-    def __init__(self):
-        self._exporter: BaseArticleItemExporter = None
-        self._master = None
-
-        self._state = None
-
-    @abc.abstractmethod
-    def setup_exporter(self, spider: BaseSpider):
-        pass
-
-    def __repr__(self):
-        return f'<{self.name} :: {self.exporter} state: {self._state}>'
-
-    @property
-    def name(self):
-        return f'{self.__class__.__name__}'
-
-    @property
-    def is_active(self) -> bool:
-        return bool(self._state)
-
-    @is_active.setter
-    def is_active(self, val: bool):
-        state = bool(val)
-        self._state = state
-        logger.info(f'{self.name} | state update: {self._state}')
-
-    @property
-    def exporter(self) -> BaseArticleItemExporter:
-        return self._exporter
-
-    @exporter.setter
-    def exporter(self, new: BaseArticleItemExporter):
-        if not isinstance(new, BaseArticleItemExporter):
-            exporter_type_msg = \
-                f'{self.name} | You are trying to set "exporter" ' \
-                f'with wrong type: {type(new)}'
-            logger.error(exporter_type_msg)
-            raise TypeError(exporter_type_msg)
-        else:
-            logger.debug(f'{self.name} | {new} exporter settled up.')
-            self._exporter = new
-
-    @property
-    def master(self):
-        return self._master
-
-    @master.setter
-    def master(self, new):
-        self._master = new
-
-    def open_spider(self, spider: BaseSpider):
-        try:
-            self.setup_exporter(spider)
-        except Exception as exc:
-            logger.exception(f'{self.name} | Error while setting up {self.name}: {exc}')
-            self.is_active = False
-        else:
-            if self.exporter is None:
-                logger.warning(f'{self.name} | "exporter" is not set up.')
-                self.is_active = False
-            else:
-                self.is_active = True
-            if self.master is None:
-                logger.warning(f'{self.name} | "master" attribute is not set.')
-
-        if self._state:
-            try:
-                self.exporter.start_exporting()
-            except Exception as exc:
-                logger.exception(f'{self.name} | Error while starting exporting with {self.exporter}: {exc}')
-                self.is_active = False
-
-    def close_spider(self, spider):
-        if self._state:
-            self.exporter.finish_exporting()
-            logger.info(
-                f'{self.name} | Successfully finished <{self.exporter.name}> '
-                f'exporter with {self.exporter.count} items exported.')
-
-    def process_item(self, item: ArticleItem, spider) -> ArticleItem:
-        if isinstance(item, ArticleItem):
-            if self._state:
-                try:
-                    self.exporter.export_item(item)
-                except Exception as exc:
-                    logger.exception(
-                        f'{self.name} | Error while exporting '
-                        f'<{item["fingerprint"]}> item: {exc}')
-        else:
-            pass
-        return item
-
-
-class GSpreadPipeline(ArticlePipeline):
-
-    def setup_exporter(self, spider: BaseSpider):
+    def setup_exporter(self, spider: BaseArticleSpider):
         if ENABLE_GSPREAD:
             self.master = GSpreadMaster(cfg.spreadsheet_title)
             self.exporter = GSpreadAIE(
@@ -147,9 +49,9 @@ class GSpreadPipeline(ArticlePipeline):
             )
 
 
-class BackupGSpreadPipeline(ArticlePipeline):
+class BackupGSpreadPipeline(BaseArticlePipeline):
 
-    def setup_exporter(self, spider: BaseSpider):
+    def setup_exporter(self, spider: BaseArticleSpider):
         if ENABLE_GSPREAD:
             self.master = GSpreadMaster(cfg.backup_spreadsheet_title)
             self.exporter = GSpreadAIE(
@@ -162,9 +64,9 @@ class BackupGSpreadPipeline(ArticlePipeline):
             )
 
 
-class SQLAlchemyPipeline(ArticlePipeline):
+class SQLAlchemyPipeline(BaseArticlePipeline):
 
-    def setup_exporter(self, spider: BaseSpider):
+    def setup_exporter(self, spider: BaseArticleSpider):
         if ENABLE_DATABASE:
             self.master = SQLAlchemyMaster(cfg.database_url, cfg.database_table_name)
             self.exporter = SQLAlchemyAIE(
